+++ src/main/java/com/powsybl/openloadflow/OuterLoopConfig.java
@@ -4,11 +4,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow.lf.outerloop;
+package com.powsybl.openloadflow;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.loadflow.LoadFlowParameters;
-import com.powsybl.openloadflow.OpenLoadFlowParameters;
+import com.powsybl.openloadflow.lf.outerloop.OuterLoop;
 import org.apache.commons.compress.utils.Lists;
 
 import java.util.List;
